+++ cmp/dataview/DataViewModel.js
@@ -7,7 +7,7 @@
 import {GridModel} from '@xh/hoist/cmp/grid';
 import {HoistModel, managed} from '@xh/hoist/core';
 import {bindable, makeObservable} from '@xh/hoist/mobx';
-import {throwIf, apiDeprecated} from '@xh/hoist/utils/js';
+import {throwIf} from '@xh/hoist/utils/js';
 import {isFunction, isNumber} from 'lodash';
 
 /**
@@ -53,7 +53,7 @@ export class DataViewModel extends HoistModel {
      *      as a function returning a StoreContextMenu. Desktop only.
      * @param {RowClassFn} [c.rowClassFn] - closure to generate CSS class names for a row.
      *      NOTE that, once added, classes will *not* be removed if the data changes.
-     *      Use `rowClassRules` instead if Record data can change across refreshes.
+     *      Use `rowClassRules` instead if StoreRecord data can change across refreshes.
      * @param {Object.<string, RowClassRuleFn>} [c.rowClassRules] - object keying CSS
      *      class names to functions determining if they should be added or removed from the row.
      *      See Ag-Grid docs on "row styles" for details.
@@ -155,16 +155,4 @@ export class DataViewModel extends HoistModel {
     clear()                         {return this.gridModel.clear()}
     setGroupBy(colIds)              {return this.gridModel.setGroupBy(colIds)}
     setSortBy(sorters)              {return this.gridModel.setSortBy(sorters)}
-
-    /** @deprecated */
-    get selection() {
-        apiDeprecated('DataViewModel.selection', {msg: 'Use selectedRecords instead', v: 'v44'});
-        return this.selectedRecords;
-    }
-
-    /** @deprecated */
-    get selectedRecordId() {
-        apiDeprecated('DataViewModel.selectedRecordId', {msg: 'Use selectedId instead', v: 'v44'});
-        return this.selectedId;
-    }
 }