+++ cmp/grid/Grid.js
@@ -19,7 +19,7 @@ import {convertIconToHtml, Icon} from '@xh/hoist/icon';
 import {computed, observer} from '@xh/hoist/mobx';
 import {wait} from '@xh/hoist/promise';
 import {filterConsecutiveMenuSeparators} from '@xh/hoist/utils/impl';
-import {apiRemoved, isDisplayed, logDebug, logWithDebug, consumeEvent} from '@xh/hoist/utils/js';
+import {isDisplayed, logDebug, logWithDebug, consumeEvent} from '@xh/hoist/utils/js';
 import {getLayoutProps} from '@xh/hoist/utils/react';
 import classNames from 'classnames';
 import {
@@ -71,13 +71,6 @@ export const [Grid, grid] = hoistCmp.withFactory({
      * @param ref
      */
     render({model, className, ...props}, ref) {
-        apiRemoved('Grid.onKeyDown', {test: props.onKeyDown, msg: 'Specify onKeyDown on the GridModel instead.', v: 'v43'});
-        apiRemoved('Grid.onRowClicked', {test: props.onRowClicked, msg: 'Specify onRowClicked on the GridModel instead.', v: 'v43'});
-        apiRemoved('Grid.onRowDoubleClicked', {test: props.onRowDoubleClicked, msg: 'Specify onRowDoubleClicked on the GridModel instead.', v: 'v43'});
-        apiRemoved('Grid.onCellClicked', {test: props.onCellClicked, msg: 'Specify onCellClicked on the GridModel instead.', v: 'v43'});
-        apiRemoved('Grid.onCellDoubleClicked', {test: props.onCellDoubleClicked, msg: 'Specify onCellDoubleClicked on the GridModel instead.', v: 'v43'});
-        apiRemoved('Grid.agOptions.rowClassRules', {test: props.agOptions?.rowClassRules, msg: 'Specify rowClassRules on the GridModel instead.', v: 'v43'});
-
         const {store, treeMode, treeStyle, colChooserModel, filterModel} = model,
             impl = useLocalModel(() => new GridLocalModel(model, props)),
             platformColChooser = XH.isMobileApp ? mobileColChooser : desktopColChooser,
@@ -527,12 +520,15 @@ class GridLocalModel extends HoistModel {
     }
 
     sizingModeReaction() {
-        const {model} = this;
+        const {model} = this,
+            {mode} = model.autosizeOptions;
+
         return {
             track: () => model.sizingMode,
             run: () => {
-                if (model.autosizeOptions.mode !== GridAutosizeMode.ON_SIZING_MODE_CHANGE) return;
-                model.autosizeAsync({showMask: true});
+                if (mode === GridAutosizeMode.MANAGED || mode === GridAutosizeMode.ON_SIZING_MODE_CHANGE) {
+                    model.autosizeAsync({showMask: true});
+                }
             }
         };
     }
@@ -649,6 +645,20 @@ class GridLocalModel extends HoistModel {
             wait().then(() => this.syncSelection());
         }
 
+        if (model.autosizeOptions.mode === GridAutosizeMode.MANAGED) {
+            // If sizingMode different to autosizeState, autosize all columns...
+            if (model.autosizeState.sizingMode !== model.sizingMode) {
+                model.autosizeAsync();
+            } else {
+                // ...otherwise, only autosize columns that are not manually sized
+                const columns = model.getLeafColumnIds().filter(colId => {
+                    const state = model.findColumn(model.columnState, colId);
+                    return state && !state.manuallySized;
+                });
+                model.autosizeAsync({columns});
+            }
+        }
+
         model.noteAgExpandStateChange();
 
         this._prevRs = newRs;
@@ -692,7 +702,11 @@ class GridLocalModel extends HoistModel {
 
     // Catches column resizing on call to autoSize API.
     onColumnResized = (ev) => {
-        if (isDisplayed(this.viewRef.current) && ev.finished && ev.source === 'autosizeColumns') {
+        if (!isDisplayed(this.viewRef.current) || !ev.finished) return;
+        if (ev.source === 'uiColumnDragged') {
+            const width = ev.columnApi.getColumnState().find(it => it.colId === ev.column.colId)?.width;
+            this.model.noteColumnManuallySized(ev.column.colId, width);
+        } else if (ev.source === 'autosizeColumns') {
             this.model.noteAgColumnStateChanged(ev.columnApi.getColumnState());
         }
     };
@@ -744,10 +758,21 @@ class GridLocalModel extends HoistModel {
         this.model.agGridModel.agApi.setFilterModel(filterState);
     }
 
-    // Underlying value for treeColumns is actually the record ID due to getDataPath() impl.
-    // Special handling here, similar to that in Column class, to extract the desired value.
-    processCellForClipboard({value, node, column}) {
-        return column.isTreeColumn ? node.data[column.field] : value;
+    processCellForClipboard = ({value, node, column}) => {
+        const {model} = this,
+            recId = node.id,
+            colId = column.colId,
+            record = isNil(recId) ? null : model.store.getById(recId, true),
+            xhColumn = isNil(colId) ? null : model.getColumn(colId);
+
+        if (!record || !xhColumn) return value;
+
+        return XH.gridExportService.getExportableValueForCell({
+            gridModel: model,
+            record,
+            column: xhColumn,
+            node
+        });
     }
 
     navigateToNextCell = (agParams) => {